me-emsc: test old vs new hyptan-smoothed
