+++ orangecontrib/spectroscopy/preprocess/me_emsc.py
@@ -92,74 +92,6 @@ def cal_ncomp(reference, wavenumbers,  explainedVarLim, alpha0, gamma):
     return numComp
 
 
-def weights_from_inflection_points_legacy(points, kappa, wavenumbers):
-    """ As ported from the original Matlab code.
-
-    Made legacy due to the following drawbacks:
-    - Only supports 3 or 4 points.
-    - Edges behave differently in 3 or 4 point cases.
-    - Transition shape depends on spacing of input wavenumbers.
-    - Essentially undefined behaviour when inflection points are too close.
-    - The wavenumbers are assumed to be equidistant.
-
-    This functions needs correctly ordered inputs. Example calls (function as f):
-    (3 points) f([3700, 2550, 1900, 0], [1, 1, 1, 1], np.arange(0, 5000, 10))
-    (4 points) f([3700, 2550, 1900, 1000], [1, 1, 1, 1], np.arange(0, 5000, 10))
-
-    For both calls the weights go towards 0 for wn > 3700. For (3 points)
-    wn < 1900 go towards 1, but for (4 points) they go towards 0 when wn < 1000.
-    """
-    # Hyperbolic tangent function
-    hypTan = lambda x_range, kap: 0.5*(np.tanh(kap*x_range) + 1)
-
-    # Calculate position of inflection points
-    p1 = np.argmin(np.abs(wavenumbers - points[2]))
-    p2 = np.argmin(np.abs(wavenumbers - points[1]))
-    p3 = np.argmin(np.abs(wavenumbers - points[0]))
-
-    # Resolution on the x-axis used in the hyperbolic tangent function
-    dx = 0.094
-
-    x1 = np.linspace(-(p1-1)*dx, 0, p1)
-    x2 = np.linspace(dx, np.floor((p2-p1)/2)*dx, np.int(np.floor((p2-p1)/2)))
-
-    xp1 = np.hstack([x1, x2])
-
-    x3 = np.linspace(-(np.ceil((p2-p1)/2)-1)*dx, 0, np.int((np.ceil((p2-p1)/2))))
-    x4 = np.linspace(dx, np.floor((p3-p2)/2)*dx, np.int(np.floor((p3-p2)/2)))
-
-    xp2 = np.hstack([x3, x4])
-
-    x5 = np.linspace(-(np.ceil((p3-p2)/2)-1)*dx, 0, np.int((np.ceil((p3-p2)/2))))
-    x6 = np.linspace(dx, (len(wavenumbers)-p3)*dx, np.int(len(wavenumbers)-p3))
-
-    xp3 = np.hstack([x5, x6])
-
-    patch1 = 1 - hypTan(xp1, kappa[2])
-    patch2 = hypTan(xp2, kappa[1])
-    patch3 = 1 - hypTan(xp3, kappa[0])
-
-    wei_X = np.hstack([patch1, patch2, patch3])
-
-    if points[3]:
-        p0 = np.argmin(np.abs(wavenumbers - points[3]))
-
-        x1a = np.linspace(-(p0-1)*dx, 0, p0)
-        x2a = np.linspace(dx, np.floor((p1-p0)/2)*dx, np.int(np.floor((p1-p0)/2)))
-        x3a = np.linspace(-(np.ceil((p1-p0)/2)-1)*dx, 0, np.int((np.ceil((p1-p0)/2))))
-
-        xp0 = np.hstack([x1a, x2a])
-        xp1 = np.hstack([x3a, x2])
-
-        patch0 = hypTan(xp0, kappa[3])
-        patch1 = 1 - hypTan(xp1, kappa[2])
-        wei_X = np.hstack([patch0, patch1, patch2, patch3])
-
-    wei_X = wei_X.reshape(1, -1)
-    dom = Orange.data.Domain([Orange.data.ContinuousVariable(name=str(float(a))) for a in wavenumbers])
-    data = Orange.data.Table.from_numpy(dom, wei_X)
-    return data
-
 
 def hyp_tan_tabulated(tol=1e-3, points=51):
     """
